+++ Goblin Camp/src/NPC.cpp
@@ -1422,7 +1422,7 @@ class NPCListener : public ITCODParserListener {
 			}
 		} else if (boost::iequals(name,"effectChances")) {
 			for (int i = 0; i < TCOD_list_size(value.list); ++i) {
-				NPC::Presets.back().attacks.back().StatusEffects()->at(i).second = (int)TCOD_list_get(value.list,i);
+				NPC::Presets.back().attacks.back().StatusEffects()->at(i).second = (intptr_t)TCOD_list_get(value.list,i);
 			}
 		} else if (boost::iequals(name,"projectile")) {
 			NPC::Presets.back().attacks.back().Projectile(Item::StringToItemType(value.s));
